Issuance - Stop sheet & page (#81)
* Introduce PinPrompt

* Animate Stepper

* Animate pages with default behaviour of AnimatedSwitcher

Experimented with implementing the 'pageview' like transition
but couldn't get it to behave properly yet. As this is not the
prio, opted for the default animation for now to make it at
least feel a bit less clunky and have an entry point of where
to start building the proper animation.

* Introduce PinPage

To make the pin entry as flexible as possible this commit
ads the PinPage. The PinPage resolves the closest PinBloc to
render it's states and communicate the inputs. The PinPage can
either be used direclty, e.g. in the issuance flow as seen in this
commit. Or it's indirectly used by e.g. the PinScreen or PinPrompt
to render the Pin Entry widgets.

* Update MockWalletRepositoryTests

Make sure the leftOverPinAttempts are handled
properly when the user attempts to confirm a
transaction.

* Add pin localizations

* Fix analyze issues

* Implement Issuance 'Stop' Sheet

* Generalize Terminal Page

Setup a terminal page widget that's to be used
to render the terminal error pages of the issuance
flow.

* Fix merging issue

Wrong line was selected in the _onIssuanceBackPressed
method that caused the back animation to behave incorrectly.