Update Onboarding to Include Checking Individual Cards (#139)
* Introduce and Support WalletPersonalizeCheckCardPage

Allow user to check data on individual cards which are
offered during the personalization phase.

* Update WalletPersonalizeCheckCardPage

Update the WalletPersonalizeCheckCardPage and
WalletPersonalizeCheckCardPage so that it renders
the Card at the top of the screen.

* Localize ExitSheet and CheckCardPage

* Introduce WalletPersonalizeConfirmPinPage

* Localize WalletPersonalizeConfirmPinPage

* Fix CheckDataOfferingPage Spacing Issue

Spacing above the top divider was uneven
when it was shown.

* Improve CheckDataOfferingPage

Delegate the building (or not) of the footer and
card to their respective methods.

* Rename a WalletPersonalize Event

* Update WalletPersonalizeCheckCards

Make it a little easier to comprehend how the
progress is calculated.

* PR Feedback

* Fix Card Ordering

The GetWalletCardsUseCase was ordering the (in memory)
list of cards. Causing the order to change, even on the
dashboard. Now removing the sorting and just returning a
copy of the list, to avoid unwanted manipulation.