* Update dependencies
* Refactor build script

* Generate `Info-Debug.plist` via build script

Adds URL schemes (`Info-Debug-Additions.plist`)
 to clean copy of `Info.plist` in `Info-Debug.plist` file

* Update plist content order (by Plistbuddy)

* Move URLTypes to separate Plist

* Add (debug) URL scheme as User-Defined Setting

* Move Intent Filter

Move the intent filter to the debug AndroidManifest

* Pull up NavigatorState

Pull up the GlobalKey that contains the
NavigatorState to make the dependency more
explicit.

* Update Info.plist Formatting

Save the file through XCode so that the formatting
is standardized.

* Make DecodeDeeplinkUseCase Synchronous

This function was using the async keyword
and returning a future, but it can actually
be synchronous, making it easier to call.

* Add iOS Deeplink Support

Define the walletdebuginteraction scheme so that
the deeplinks also open the app on iOS.

Can be tested on the simulator with:
`/usr/bin/xcrun simctl openurl booted "walletdebuginteraction://debug#%7B%22id%22%3A%20%22HEALTH_INSURANCE%22%2C%22type%22%3A%20%22issue%22%7D"`

* Introduce DoOnInit Widget

Instead of calling processQueue() on every
HomeScreen rebuild we now only call it when
the HomeScreen route is being created through
the initState callback of the new DoOnInit widget.

* Introduce DeeplinkService & Dependencies

Create a central place to handle incoming deeplinks
in the form of a deeplink_service.dart. This class
observes the initial and incoming deeplinks and
handles them by decoding them and triggering the
desired navigation. It also makes sure the app is
ready to handle the navigation events before triggering
them, as they should be queued when the wallet isn't
configured yet

* Add uni_link and update AndroidManifest.xml

- Add the uni_link dependency to easily get notified
about incoming deeplinks.
- Update the AndroidManifest so that deeplinks with the
walletdebuginteraction scheme are forwarded to the app.