Update Theming
* Fix Dart Analyze Issues

* Update CI Images

Update reference to the CI's docker images
to those that use Flutter 3.7.1

* Update CI Images

Update reference to the CI's docker images
to those that use Flutter 3.7.1

* Move Dark/Light Widget Themes

Opted to move the color alteration of the
base themes to the dedicated Light/DarkWalletTheme
classes.

* Force use of LightTheme

* Setup BaseWalletTheme

Create a BaseWalletTheme which contains all
theme styling except for the light/dark specific
coloring. The actual coloring is applied in the
wallet_theme using the colors provided by the
respective light/dark theme classes.

* Split out divider/bottomSheet theme

Start splitting out themeData, making it easier
to keep the basics while adjusting the colors to
support future dark theming.

* Update bottomSheetTheme

Update the bottomSheetTheme to make sure
the bgColor and shape is correct when M3
is enabled.

* Disable AppBar Elevation Effect

With the new M3 spec the app bar changes
color when content is scrolled underneath
(the new elevation feedback), disable this
for now by setting the scrolledUnderElevation
to 0, as this new color changing effect is
inconsistent with the Figma designs.

* Update BottomSheetDragHandle color

Update the color of the BottomSheetDragHandle
to match the latest Figma designs at #445581.

* Fix MockDigidScreen Theme Override

With M3 enabled the elevatedButtonTheme
needs to be overridden as well to set an
explicit backgroundColor for the bgColor.

* Update textButtonTheme to work with M3

Update the shape of the textButtonTheme so that
the ripple behaves as expected when enabling M3.

* Update elevatedButtonTheme to work with M3

Also noticed radius was set to 8 instead of 12
which is shown in Figma.

* Make indicatorColor equal primary color

Update the indicatorColor to rely on the primary
color directly and update the tabBarTheme to make
sure the TabBar looks consistent when enabling
material3 styling.

* Move primaryColor to colorScheme only

Instead of relying on the primaryColor from the
ThemeData directly, move it to the colorScheme
and reference it from there.

* Update dividerStyle

Set a dedicated dividerTheme and align
it with the outlineVariant color.

* Rely on Theme.of(context).colorScheme.error

Instead of using a mix of Theme.of(context).errorColor
and Theme.of(context).colorScheme.error, rely on the
errorColor from the colorScheme consistently.

* Move fontFamily

Pull up the fontFamily constant as it will be
used in both the light and (future) dark theme.

* Fix Deprecation Warnings

Fix all deprecation warnings and update the
ThemeScreen to reflect the new textTheme naming.

* Fix Compile Issues

* Fix iOS Build

Reset / Update Pods to make sure the iOS
app builds as expected.