Integrate flutter_rust_bridge
* Move Generated Rust Bridge Code

Move the generated bridge code to the a
'bridge_generated' subdirectory.

* Update rust_core/.gitignore

* Update build.gradle

- Remove x86 target
- Remove rustup x86 requirement
- Reformat build.gradle
- Remove some excess documentation

* chore: Bump Rust testing image

* Add manual build step for Rust compile to Xcode project instead of relying on cargo-xcode

* ci: Test and build Rust code

* Update .gitignore to ignore entire jniLibs dir and Xcode project user files in Rust core

* Improve README.md

Add missing documentation about cargo-ndk and
flutter_rust_bridge_codegen crates. Also document
how to generate the bridge code.

* Update README.md

Update the setup development requirements
section to include the new Rust requirements.

* Implement pin validation

* Pass vector parameters as slice, extract constant and clippy fixes

* Implement pin validation

The following validation rules are implemented:
- pin must have length 6
- pin must only contain digits
- pin must contain at least 2 unique digits
- pin may not be ascending or descending

* Ignore rust_core generated libs

* Remove .so files

* Validate Pin in Rust

Swap out the dart implementation for the Rust
variation to handle pin validation.

* Setup Flutter iOS <-> Rust Bridge

Configure the Xcode Runner project to include
the native rust_core code, allowing it to be
called when running on iOS as well.

* Fix Flutter Lint Issue

* Simplify Rust Bridge

Only expose an 'isValidPin' api for now

* Integrate flutter_rust_bridge

This commit performs the basic Android setup to
use the flutter_rust_bridge package.

Note that besides the new code your local
system also has to be configured to compile
the rust code. This setup includes:
- Installing rust
- Installing the NDK
- Using rustup to add the targets
(aarch64-linux-android, armv7-linux-androideabi,
x86_64-linux-android, i686-linux-android)
- Configuring the corresponding linkers (~/.cargo/config)
- Adding the ANDROID_NDK to your gradle.properties
- Installing cargo-ndk & flutter_rust_bridge_codegen through cargo

As a sample 3 methods are added to the api.rs on the rust side:
- hello_world() -> Sample method that simply returns a String
- get_card(id) -> Sample method that takes a simple parameter
and returns a custom object defined in rust.
- set_card(card) -> Sample method that passes an object to rust
and reads one of its fields.

This should proof the initial viability of this package.

Furthermore the app/build.gradle has been updated to compile the
rust code automatically when the app is launched, this includes
copying in the new .so files for each architecture.
When the api.rs is updated, it is currently the responsibility
of the developer to re-run `flutter_rust_bridge_codegen` to
generate the new bridge sources.

For more info on the required setup, see:
https://cjycode.com/flutter_rust_bridge/integrate.html