Improve PinSetup Accessibility
* Tweaks 'fits logo' logic

With the PinField now inside the top expanded
bloc the _requiredHeightToShowLogo had to be
tweaked. Now also taking font scaling into account.

* Rename generic `TestWidget`

Rename it to WalletAppTestWidget, indicating
it contains/provides wallet related data, like
the theme and translations.

* Fix Backspace Announcements

When pressing backspace while setting up a pin the
announcements were not consistent. Fixed by providing
this knowledge to the screen and acting on it
accordingly.

Also fixed some localization issues.

* SetupSecurityPinPage Layout Tweaks

Improve the layout of the SetupSecurityPinPage to
accomodate the error message on a device with font
and display scaling set to large, while still
maintaining the desired design for normal devices.

* Tweak Accessibility Announcements

Tweak the announcements so that they work more
consistently.

Also localize the tapHint of the digit key.

* Sort ARB files

* Provide Localizations to TestWidgets

* Localise Accessibility Texts

* Add PinKeyboardTest

Add some basic tests for the PinKeyboard to
validate it's behaviour and contrast.

* Announce Setup Pin Page States

The PinKeyboard maintains focus when going to the confirmation
step, and focus in accessibility can't be forced. As such opted
to announce important page changes so that the user is aware
of the required action.

* Add Accessibility Hints to PinButtons

- Notify the user how many digits have been entered
when focusing on the pin field
- Notify the user of the behaviour of the backspace
button
- Announce keys as 'entering' a digit instead of the
default 'activate'