Allow User Language Selection
* Fix PR Feedback

* Update WalletBlocProvider Docs

* Add LanguageRepository Tests

* Fix Import Issue

* Refactor LanguageTag

Rename to LanguageCode to be more consistent
with the Locale api.

* Provide Language Labels

Instead of showing the languageCode
in the ui, provide the language's
friendly name instead. Done by keeping
a map containing these values direclty
inside the BloC, as these values are static
and not supposed to change with the selected
locale.

* Rename Some Language Classes

- Rename DataSource -> Store, as it only persists
a setting.
- Rename LanguageCubit -> PreferredLocaleCubit

Also provide the default locale to the ChangeLanguageBloc
if no override locale has been selected, to be able to highlight
the correct item.

* Setup Preferred Language Persistence

Setup LanguageRepository, Source and Cubit
to persist the user preferred language and
reflect this in the app.

* Add Navigation to ChangeLanguageScreen

- From introduction
- From settings

* Setup ChangeLanguageScreen

- Setup Screen
- Setup Basic Bloc
- Setup Route