Support Landscape in Introduction
* Fix Landscape Image Sizing

Image did not always fill the full height of
the screen.

* Simplify IntroductionPage

* Add Basic Intro Landscape Support

* Disable Portrait Lock

* Fix ScreenReader Issues

- Screenreader announced invisible (space placeholder)
button
- Screenreader read part of time because semantics
wrapped the SafeArea which could (partly) include
the clock on some devices
- Order of the screenreader's traversal tweaked, so
the back button is not centered between the screen
text and action buttons.

* Clean up LanguageButton definition

Avoid specifying it two times

* Specify 'back' as WCAG key

Create a dedicated WCAG/Tooltip translation
for the back action.

* Update Introduction Secondary Button

- In dev mode also display the language button
- Ignore the secondary button when it's hidden
(before this change the screenreader would still
highlight it, even while it wasn't drawn)

* Update Introduction Translations

* Fix TextScaleFactor

Supply a fixed textscalefactor to the
headline text of the introduction_page.

Approved from an accessibility perspective
because un unscaled headline is still larger
than a scaled body text.

Doing so allows the UI to not get overly cramped
or even overlap with the paging indicator.

* Tweak Intro Back Button

Tweak the button so that:
- The size meets minimum accessibility requirements
- Add semantics so it can be announced
- It uses the theme's colors
- Make sure it's ignored for talkback when it's hidden