Provide Specific Pin Error Messages
* Update Default MOCK_REPOSITORIES value

Default to true

* PR Feedback

WalletRepositoryImpl -> CoreWalletRepository

* Update PinUseCaseTest

Check for explicit error.

* Extract PinValidationErrorMapper

Extract the mapper and inject it into the repository,
making things easier to swap out and test.

* Update wallet_repository_impl

Be consistent in throwing the UnimplementedError.

* Fix check_is_valid_pin_usecase_test issues

* Accessibility Fix

When the error message becomes multi line,
the text can still overflow (pixel 2, large font &
display scaling enabled). To avoid this we disable
the logo's bottom spacing when it's invisible
(since the logo has top and bottom spacing this
bottom spacing is excessive when the logo is
hidden). Doing so allows 2 line errors to fit
on the configuration described above.

* Update CheckIsValidPinUseCase

Update the CheckIsValidPinUseCase so that it's
guaranteed to only throw a PinValidationError.

* Add basic CheckIsValidPinUseCase Test

* Remove root rust_core

Since it's unlikely we'll be building a web
app based on this project soon or ever, removing
the rust_core.dart file from the root of the
repository to have less ambiguous imports.

* Restructure Mocks/Impls

Be consistent in where we place the
mocks and implementations of the
repositories. Previously it was a mix
of in the root of the corresponding repo
directory and in the mock/impl subdirs.
Now we consistently use the subdir approach.

* Separate Repository Mocks and Impls

Allow selection of mocks vs implementations
of the repositories when building the app.
This can be done by providing the following build
argument: --dart-define MOCK_REPOSITORIES=true

Defaults to building with normal implementations, note
that these are currently 99% unimplemented, making it very
obvious what is yet to be done.

Note that there are Repositories whose implementations
might be complete, but they rely on memory or mocked
datasources, since there is no Mock for these Repositories
these are thus not provided. It might eventually be useful
to also provide a mock for these repositories, but for now
that doesn't seem to be the case.

* Map PinValidationErrors

Map the PinResult to a local PinValidationError
that bubbles up to the UI through the corresponding
state class.

* Move RustCore Dependency

Avoid injecting the RustCore directly into
a UseCase but keep the Repository abstraction
layer.

Still WIP as a decision needs to be made on who
is responsible for doing the bincodeDeserialize()
for the SERDE generated models.

* Update analyzer rules

Update the excluded directory.

* Update imports

* Refactor 'core' package

Move it to /pub/core_domain, the idea being
that all our local dart packages can live in
the /pub directory to avoid bloating the root
of the project.

Rename it to `core_domain` to be a bit more specific
about the origin of these models.

* Update README.md

- Fix typo in documentation
- Elaborate on what is generated and where

* Update rust_core README.md

* Move Generated Rust Bridge Code

Move the generated bridge code to the a
'bridge_generated' subdirectory.

* Update rust_core/.gitignore

* Update build.gradle

- Remove x86 target
- Remove rustup x86 requirement
- Reformat build.gradle
- Remove some excess documentation

* Fix Lint Warnings

Fix lint warnings in the serialization dependencies,
as they are not generated I opted to fix these
instead of adding them to the ignore list.

* Fix PinErrorText Alignment

When text spans multiple lines it became
apparent that the text is not centered, fixed
by setting the textAlign.

* Exclude Domain from Analyzer

Exclude the generated domain classes from the
flutter analyzer, as they do not conform to our
linting rules.

* Propagate PinError to UI

* chore: Bump Rust testing image

* Add manual build step for Rust compile to Xcode project instead of relying on cargo-xcode

* ci: Test and build Rust code

* Update .gitignore to ignore entire jniLibs dir and Xcode project user files in Rust core

* Update rust_core/README.md

Updated the location where the generated code
is placed, location changed so that it can be
easily included in the Flutter project's pubspec.

* Use c_style_enums for swift generation

* Export Generated Dart code to 'domain' package

Export the models to a domain package and use
these models to deserialize response of the
isPinValid bridge method call.

* Update Generator Config

Set with_c_style_enums to true to
output cleaner dart models.

* Improve README.md

Add missing documentation about cargo-ndk and
flutter_rust_bridge_codegen crates. Also document
how to generate the bridge code.

* Update README.md

Update the setup development requirements
section to include the new Rust requirements.

* Use bincode for data type serialization

* Implement pin validation

* Pass vector parameters as slice, extract constant and clippy fixes

* Implement pin validation

The following validation rules are implemented:
- pin must have length 6
- pin must only contain digits
- pin must contain at least 2 unique digits
- pin may not be ascending or descending

* Ignore rust_core generated libs

* Remove .so files

* Validate Pin in Rust

Swap out the dart implementation for the Rust
variation to handle pin validation.

* Setup Flutter iOS <-> Rust Bridge

Configure the Xcode Runner project to include
the native rust_core code, allowing it to be
called when running on iOS as well.

* Fix Flutter Lint Issue

* Simplify Rust Bridge

Only expose an 'isValidPin' api for now

* Integrate flutter_rust_bridge

This commit performs the basic Android setup to
use the flutter_rust_bridge package.

Note that besides the new code your local
system also has to be configured to compile
the rust code. This setup includes:
- Installing rust
- Installing the NDK
- Using rustup to add the targets
(aarch64-linux-android, armv7-linux-androideabi,
x86_64-linux-android, i686-linux-android)
- Configuring the corresponding linkers (~/.cargo/config)
- Adding the ANDROID_NDK to your gradle.properties
- Installing cargo-ndk & flutter_rust_bridge_codegen through cargo

As a sample 3 methods are added to the api.rs on the rust side:
- hello_world() -> Sample method that simply returns a String
- get_card(id) -> Sample method that takes a simple parameter
and returns a custom object defined in rust.
- set_card(card) -> Sample method that passes an object to rust
and reads one of its fields.

This should proof the initial viability of this package.

Furthermore the app/build.gradle has been updated to compile the
rust code automatically when the app is launched, this includes
copying in the new .so files for each architecture.
When the api.rs is updated, it is currently the responsibility
of the developer to re-run `flutter_rust_bridge_codegen` to
generate the new bridge sources.

For more info on the required setup, see:
https://cjycode.com/flutter_rust_bridge/integrate.html