Proposal: Introduce Screenshot Testing
* PR Feedback

Add some documentation to indicate this check is
a temporary solution.

* Update SetupSecurity Golden Test

Verify the most diverse states:
- Setup in progress
- Validation feedback
- Setup failed

* Update SetupSecurity Goldens

The 'landscape' support merge for the pin
screen caused the layout to be updated, and
therefor the golden tests related to this to
fail. Normally this update should be part
of the 'pin keyboard landscape support' MR
but since the screenshot tests aren't merged
yet it has to be done after merging in main.

* Fix test documentation

Inverted the commands that describe how to verify
the goldens or exclude them when running tests.

* Update CocoaPods lockfile

* Update README.md

Update the readme to include some info on the
UI tests.

* Document withState method

* Add Initial SetupSecurityScreen UI Test

The order in which the scenarios are rendered
and built cause the widget to always be rendered
with the same (last set) state for a bloc when
relying on the `onCreate` of the addScenario
method to provide a state. Since the bloc is
provided higher up in the tree. To fix this
and allow us to test the screen with easily
configurable states I introduces the `withState`
extension, which takes a bloc and the desired
state. This way the provider is part of the running
scenario, and thus only used for that scenario.

* Remove navigatorObservers

Unused for now, might be useful te re-introduce
when we want to test for navigation events.

* Make ui-test-failure artifact conditional

Only add this artifact when the screenshot
tests result in a failure.

* Add test-flutter-ui job

Add a (manual) job that runs the
screenshot (golden) tests.

* Disable golden tests on CI

* Revert "Split up Goldens"

This reverts commit 6fe19780a112df9e45bb604dec524736b5f840ec.

* Split up Goldens

- Split up the screenshots that represent the
golden standard for the screenshot test, in
an attempt to minimize the diff between the
CI and the developer's local machine
- Introduce an acceptable diff threshold of
0.5% to see if this is enough for the CI to
agree with the locally golden standard.

* Disable 'Dev' button for test in intro

* Disable Golden Test Fonts

Temporarily disable the usage of fonts
on for the golden tests to check if that
is the sole cause of the test issues.

* Fix dart analyze issue

* Setup DeviceUtils

Refactor the GoldenToolkitUtils class a little
bit so that, making it more specific to providing
test Device related classes.

Also fix a broken PinKeyboard test.

* Add IntroductionScreen Test

Provide a test that verifies all
steps of the introduction screen.

* Combine Scaling Test

* Golden Test Setup

Add dependency and provide initial
sample setup for the splashscreen.