Update Personalization Flow UI
* Update DigidConfirmAppPage Parameters

Rename the parameters to be more consistent
in on button press callback naming.

* Fix scaling related issue

Instead of spanning 2 lines the text would
overflow without the expanded widget.

* Clean up WalletPersonalizeBloc

Clean up the bloc and update the progress
stepper states.

* Fix lint issues

* Map PidAttributes

Add a mapper to map the pid attributes so that
they meet the new UI requirments (i.e. rendering
them together and with a dedicated icon).

* Introduce WalletPersonalizeDigidErrorPage

Add a placeholder page for when the user
cancels out of the DigiD flow.

* Introduce MockDigid Error Path

Add the possibility to decline the wallet
during the digid login process so that the
unhappy flow can be simulated.

* Sort ARBs

* Introduce WalletPersonalizeNoDigidPage

* Update WalletPersonalizeIntroPage

Update the text alignment of the
WalletPersonalizeIntroPage to match
the new design.

* Restructure WalletPersonalizeCheckDataOfferingPage

Update the WalletPersonalizeCheckDataOfferingPage
to be self contained and start rendering icons.

* Remove Photo from PID

* Clean up WalletPersonalization Flow

Remove all screens related to adding extra
government cards, as they are obsolete for now.