Android Keystore ECDSA Support
* Only add integration-test feature on debug builds

Make the addition of the integration-test optional,
so that it is only included when running debug builds.
The connectedAndroidTest which relies on this always
builds the app in debug mode so will still succeed.

* Fix typo

Remove a space that shouldn't be there

* Call hw_keystore_test_hardware_signature from Android Instrumented Test

Paired with Daan to setup an instrumented test that calls into rust
and in term calls the `hw_keystore_test_hardware_signature` method to
align with the iOS integration test.

* Updated HwKeyStore Instrumented Test

* Update clean tasks

Make sure the generated .so files and
bindings are cleaned when running the
clean task.

* Add Auto Initialization

Use the `startup-runtime` library to make initialization
of the library seamless. This library wraps the initialization
in a content provider, which is guaranteed to always be launched
when the app is started. This way the consumer of this module, in
our case the flutter app, does not need to call any init method
as this is done automatically through this content provider.

* Add Android ECDSA Support