Add ECDSA Tests & KeyExceptions
* Add `workingDir` variable

* Define variables for JNI + bindings directories

* Clean up Gradle syntax

* Store main project JNI directory for reuse

* Update docs to run integration tests

* Improve isRunningOnEmulator check

* Revert uni_links dependency change

pubspec.lock was committed with experimental
uni_link dependency, reverting.

* Enforce Hardware Backed Keys

The generated keys should always be enforced
to be hardware backed, if they are not an
exception will be thrown. The only situation
where we do allow software backed keys is when:
- The app is running in debug mode
AND
- The app is running on an emulator

This allows us to run integration tests on
emulators (which do not have hardware backed keys).

* Fix platform_support integration test

Re-introduce the espresso dependency to allow
the integration test to run correctly.

* Update Dependencies

- Update the gradle wrapper
- Update the android build tools

* Pin IRMAMobile reference

Pin the reference to a IRMAMobile source file
to a specific repository tag.

* Add initial proguard-rules

Configure proguard so that the initializer
is not stripped when compiling in release mode.

* Remove platform_support srcDirs addition

Removed the explicit source dirs, as they should be added
automatically through the platform_support module. This
was causing clashes when compiling in release mode.

* Remove unused generateKey return value

The public key that was returned was not used,
and could thus be removed.

* Throw KeyExceptions

Throw the exceptions as defined in the bridge
udl file.

* Introduce ECDSAKey Tests

Add some basic tests for the ECDSA
keystore implementation.