Override loaded .so Library
The app was crashing because the platform_support
library was always trying to load the libplatform_support.so
file, even though all this code is part of the librust_core.so
file and therefor the libplatform_support.so file is excluded
when packaging the app to safe space. This commit allows
consumers of the platform_support module to override which
library file should be loaded by setting some metadata in
the AndroidManifest.xml.