Bubble up Key securityLevel
Make sure we log the securityLevel in the
exception so that we can track (when error
reporting is introduced) if the
`SECURITY_LEVEL_UNKNOWN_SECURE` is ever
reported and thus should potentially be
treated differently.

Note that securityLevel is only supported on
android >=31, as such I defined a generic
'unsupported' code which I set to a very
unlikely -999 value.