Remove Duplicate Android Code
Instead of checking if the device is
unlocked inside the key generation method,
pull it up a level into the HwKeyStoreBridge
to avoid the duplicated code and make sure we
verify it whenever a key is looked up.

Also add strongbox extension method to clean
up the api and remove the duplication of the
platform/feature check.