Add database crate
Add a database crate with the initial (WIP)
database setup. With this commit a database
is created at a static location, to which
we write a simple model and verify that this
happens correctly. Additionally there is a
password generator which creates a password
if none is set, and otherwise returns the
previously created password.

This password and db behaviour is verified
in some simple rust tests.