Introduce TermsAndConditionsScreen
* Remove unneeded backbutton

As it's added by default when the route
is poppable.

* Remove inferred BackButton

It is automatically added when the route is
poppable, so removing the explicit mention of
the backbutton.

* Remove unused import

* Update formatting

* Sort ARBs

* Add TermsAndConditionsScreen Localizations

* Add DiscoverMoreSection

* Introduce TermsAndConditionsScreen

Add the TermsAndConditionsScreen and an
ExtendedPolicyRow to render the policy entries.

* Update method names

Remove final detailS reference.

* Process PR Feedback

Rename things: Details -> Detail

* Rename CheckAttributesDetail -> CheckAttributes

Removed the 'Detail' qualifier as I believe the
name 'CheckAttributes' actually covers the correct
load by itself.

* Add Accessibility Scrollbar

Always show the scrollbar if content
can be scrolled.

* Rename CheckAttributeDetails -> CheckAttributesDetail

More consistent with other 'Detail' screens/pages

* Update _buildConditionsRow return type

Return more generic Widget instead of
an InfoRow.

* Remove unused import

* Add ScrollBar

Make sure the scrollbar is always visible on the
organization detail screen (for consistency and
accessibility reasons).

* Refactor ApproveOrganizationPage -> OrganizationApprovePage

Be more consistent with OrganizationDetailScreen and
the corresponding package structure.

* Update OrganizationDetailScreen Formatting

* Move ApproveOrganizationPage

* Refactor OrganizationDetails -> OrganizationDetail

* Sort ARBs

* Introduce CheckAttributeDetailsScreen

Add the CheckAttributeDetailsScreen, used
to let the user verify the attributes she
is about to share in the verification process.

* Update confirmButton icon

* Update Policy Icon

* Add Missing Organization Title

* Add localizations

* Add CardAttributeRow to DesignSystem

* Update Verification Attributes Screen

- Update the BLoC so that it also provides
the cards from which the attributes were
fetched, as this is needed in the new ui.
- Introduce CardAttributesRow, widget that
shows which attributes are includes from a
specific card.

* Introduce InfoRow

Introduce an InfoRow Widget to render the common
ui of showing a row with an icon, title, subtitle
and optional chevron (when onTap is provided).

* Update VerificationConfirmDataAttributesPage Title

Update the title to include the count of the to
be shared attributes.

* Fix flutter analyze issue

* Fix localizations

- Remove !?
- Update english translations

* Update Mock

Description contains amsterdam so
better align the mock data for the
sake of consistency.

* Sort ARB files

* Add localizations

* Make info fields conditional

* Update OrganizationDetails States

- Add 'firstInteraction' logic
- Render Failure State with retry
button

* Sticky BottomBackButton

- Make the bottom back button on the
organization details screen sticky.
- Remove the BottomBackButton rounded
corners
- Add a .show() method to the
OrganizationDetailsScreen as a more
convenient api to simply show the screen
from a flow. Might also come in useful later
when we want to return some kind of result
from this screen (e.g. data incorrect or
abort request).

* Rename organization_details_screen file

* Add Privacy & Info Section Links

Make certain pieces of texts act as
hyperlinks, currently redirecting to
the PlaceholderScreen.

* Add Static Content

* Add Section Placeholders

* Configure OrganizationDetailsBloc

* Setup OrganizationDetailsScreen

* Rename OrganizationButton -> OrganizationRow

* Fix Flutter Analyze Issues

* Update ic_first_share icon

* Remove Unused IconPosition Enum

* Update Localizations

Add the newly required localizations to
the arb files and use them in the organization
approval widgets.

* Dynamic 'first interaction' flag

Make the 'eerste deelverzoek' banner dynamic,
meaning it is now based on actual historic
interactions by the app.

* Update ApproveOrganizationPage Docs

- Add some documentation to the parameters
- Optionally render the 'eerste deelverzoek' row

* Update ApproveOrganizationPage UI

* Remove Bottom Dividers

* Restructure Common Widgets

Move some common widgets that can
easily be grouped into their relevant
nested directory.

* Update OtherStylesTab

Add missing widgets to the overview

* Restructure ThemeScreen

Created dedicated files for the
tab related widgets.

* Introduce IconRow

* Update ThemeScreen Buttons Tab

* Update ThemeScreen

Update theme_screen to be a more
complete representation of all the
available widgets throughout the app.