Move signing key trait mocks in wallet_common
Move the mock implementation of EcdsaKey, EphemeralEcdsaKey and
SecureEcdsaKey by SigningKey back to the file where the traits are
defined. The reason for this is that these mock implementations are used
by multiple tests in two distinct files.

This also adds a "mock" feature to wallet_common, so that other crates
may use this mock in their tests.