Have Wallet use actual RemoteAccountServer client
This moves the AccountServerClient trait from wallet_common to the
wallet crate, since it is an abstraction of what the Wallet type needs.
The implementation of AccountServerClient by AccountServer is also moved
to the wallet crate by importing wallet_provider as a dev dependency.

Note that the account server URL and public key are now hardcoded within
the wallet crate. This is untested for now.