Replace init_async with init_wallet in flutter_api
The new init_wallet() combines initialization of the async runtime and
initialization of the wallet itself, which can now be fallible. It
replaces the static Lazy instance of the wallet with a static OnceCell.

This also removes the is_registered() API function, as it is now
integrated in init_wallet().