Simplify Wallet by adding fallible initializer
Replace the public new() constructor and load_registration() methods
with a new fallible init() method, now that flutter_api supports
initializing the wallet in a fallible manner. This also gets rid of the
added complexity of the RegistrationData enum, as the Registration data
will always be loaded during the lifetime of a Wallet instance.