Update HistoryDetails Screen
* Update Detail Header

Use the organization row as the new
header and allow users to click on
the header to navigate to the
organization detail page.

* Add Dedicated TimeFormatter

Introduce a HistoryDetailsTimeFormatter
to make sure time is formatted as per
design.

* Sort ARBs

* Localize dataAttributeSectionTitle

* Padding Tweak

Fix an issue with the padding above
the data attributes section, making sure
it also stays consistent when looking
at the 'sign document' page.

* Update Localization

* Introduce SliverDivider

Clean up the addition of divider slivers
a little bit by introducing a `SliverDivider()`
that replaces the previous definition of:
`SliverToBoxAdapter(child: Divider(height: 1)`

* Simplify the DataAttributeSection Loop

Update the HistoryDetailLoadSuccess state
so that it directly provides a map of WalletCards
to List<DataAttribute>, so that we don't have to deal
with a nullable call to `cardById`.

* Add InfoRows

Add the policy, problem and helpdesk
info row widgets to the bottom of the
HistoryDetailScreen.

* Align ContractSection

Update the padding provided to the
DocumentSection so that it aligns
properly on the HistoryDetailScreen.

* Update HimelineAttributeRow

Update the left padding of the
HistoryDetailTimelineAttributeRow
so that it aligns properly with the
DataAttributes (and their new spacing)

* Update OrganizationLogo Size

* Show SourceCard Title

Update the HistoryDetail Screen
and Bloc to provide the cards so
that the title can be shown on in
the DataAttributeSection.

* Update DataAttributes Layout

Update the layout to include a mention
of the source card. Currently showing
the id, to be migrated to the card's
name.

* Update Interaction Title