Introduce ForgotPinScreen
* Update PlaceholderScreen

Set disable the secured flag, so that
the placeholder screen is not covered by
another pin screen when the user has
forgotten her pin.

* Localize ForgotPinScreen

* Introduce ForgotPinScreen

This screen is shown when the user
indicates she forgot her pin, it should
help to evaluate the possible recovery
options.

* Rename WalletPersonalizeNoDigidPage

Rename it to WalletPersonalizeNoDigidScreen as
it represents a full screen which we can navigate
to, instead of a page that is part of the flow.