Update CheckAttributesScreen
* Move Navigator.pop() call

Instead letting the previous (hidden) screen be
in charge of popping it's child routes,
this popping behaviour is now pulled up, into
the currently visible screen. As soon as an
onDataIncorrectPressed or similar callback is
triggered we now make the assumption that the
user is done interacting with this page and thus
it's safe to be popped. Imo this results in a
slightly more consistent, cleaner api.

* Handle onDataIncorrectPressed

Instead of moving to the placeholder
screen we now navigate to the report
issue screen when the user indicates
something is not right when reviewing
her attributes.

* Update CheckAttributesScreen

Update the CheckAttributesScreen so that
it shows the title of the attributes
source card.